added test + remove \n from string - need to update detect-secret vetsion
